refactor(Lotto): 파일 이동 금지 규칙을 지키기 위해 Lotto 파일의 위치 원래대로 이동
